refactor(tui): split model_tree.go into focused files (#250)
The 726-line model_tree.go file has been refactored into 5 focused, single-responsibility files:
- model_tree_view.go: View modes, visibility, cursor selection
- model_tree_expansion.go: Expansion/collapse, cursor movement, state management
- model_tree_build.go: Tree building, filtering, path operations
- model_tree_navigation.go: Node identification, cursor restoration, name resolution
- model_tree_helpers.go: Test helpers, group utilities

This improves code maintainability and makes each concern easier to understand and modify.

Changes:
- Removed dead code: pruneEmptyGroups and other unused functions (22 lines)
- Removed model_tree.go linter exclusion from scripts/lint.sh
- Fixed formatting error in internal/format/notification.go (uppercase to lowercase)
